pyproject.toml
[tool.poetry]
name = "weed-fucker-5000"
version = "0.1.0"
description = ""
authors = ["Dustin Maselbas <dmaselbas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.19"
esptool = "^4.7.0"
docopt = "^0.6.2"
opencv-python-headless = "^4.10.0.82"
mypy = "^1.10.0"
pigpio = "^1.78"
rplidar = "^0.9.2"
paho-mqtt = "^2.1.0"
aiortc2 = "^1.8.0.post0"
aiohttp = "^3.9.5"
opencv-python = "^4.10.0.84"
pyserial = "^3.5"
pynmea2 = "^1.19.0"
plotly = "^5.22.0"
streamlit = "^1.36.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
